fix login page

+++ lib/src/bussines_logic/bloc/home_state.dart
@@ -3,14 +3,14 @@ part of 'home_bloc.dart';
 abstract class HomeState {}
 
 final class HomeInitial extends HomeState {
-  double defaultValue;
+  int defaultValue;
   HomeInitial({this.defaultValue = 0});
 }
 
 final class HomeUpdate extends HomeState {
-  double temp;
-  double humd;
-  double waterVolume;
+  int temp;
+  int humd;
+  int waterVolume;
 
   HomeUpdate(
       {required this.temp, required this.humd, required this.waterVolume});
